build: extract readme coverage macro logic into a separate build-readme module
